+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - Add **icupy.icu.NumberFormat.ERoundingMode.ROUND_HALF_CEILING**
   - Add **icupy.icu.NumberFormat.ERoundingMode.ROUND_HALF_FLOOR**
   - Add **icupy.icu.NumberFormat.ERoundingMode.ROUND_HALF_ODD**
+  - Add **icupy.icu.UCalendarDateFields.UCAL_ORDINAL_MONTH**
 - CHANGED
   - Update [pybind11](https://github.com/pybind/pybind11) to v2.10.4
 - FIXED